+++ VirtualH89/Src/h-17-4.h
@@ -25,15 +25,15 @@ class H_17_4: public DiskDrive
     H_17_4();
     virtual ~H_17_4();
 
-    void getControlInfo(unsigned int pos, bool& hole, bool& trackZero, bool& writeProtect);
+    void getControlInfo(unsigned long pos, bool& hole, bool& trackZero, bool& writeProtect);
 
     void selectSide(BYTE side);
     void step(bool direction);
 
-    BYTE readData(unsigned int pos);
-    void writeData(unsigned int pos, BYTE data);
+    BYTE readData(unsigned long pos);
+    void writeData(unsigned long pos, BYTE data);
 
-    virtual BYTE readSectorData(BYTE sector, unsigned int pos);
+    virtual BYTE readSectorData(BYTE sector, unsigned long pos);
     void insertDisk(FloppyDisk* disk);
 
   private:
